Transver ImageRenderWrapper in ImageRander.ts file
